# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Healthy App",
        "description": "API documentation for Healthy",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get a list of orders",
                "operationId": "2cd7e03021e72d9d799c12dfe179c865",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "query",
                        "description": "Filter by user ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by status",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "fromDate",
                        "in": "query",
                        "description": "Filter by start date",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "toDate",
                        "in": "query",
                        "description": "Filter by end date",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/OrderResource"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Create a new order",
                "operationId": "d690d6834318e9b064df441e854de8b9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreOrderRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Order created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/orders/user-self": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get a list of orders",
                "operationId": "4ac6f21a2d2099e636ca1790c49f73bb",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "query",
                        "description": "Filter by user ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by status",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "fromDate",
                        "in": "query",
                        "description": "Filter by start date",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "toDate",
                        "in": "query",
                        "description": "Filter by end date",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/OrderResource"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/orders/{id}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get a specific order",
                "operationId": "c29621029e70f65247909d3e09031118",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the order to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Orders"
                ],
                "summary": "Update an existing order",
                "operationId": "6a7029f5510ca08e28a2d58a5fe19439",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the order to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateOrderRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Order updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Orders"
                ],
                "summary": "Delete an order",
                "operationId": "303a7b797887b66c7790c4c187b54790",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the order to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order deleted successfully"
                    },
                    "422": {
                        "description": "Only pending orders can be deleted"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/orders/status-options": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get order status options",
                "operationId": "fd6b0d74b6ce84b1af55b9b508881e24",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreOrderRequest": {
                "required": [
                    "userId",
                    "deliveryAddress",
                    "deliveryTimeSlot",
                    "items"
                ],
                "properties": {
                    "userId": {
                        "description": "The ID of the user placing the order",
                        "type": "integer",
                        "example": 1
                    },
                    "deliveryAddress": {
                        "description": "The delivery address for the order",
                        "type": "string",
                        "example": "123 Main St, Springfield, IL"
                    },
                    "deliveryTimeSlot": {
                        "description": "The selected delivery time slot",
                        "type": "string",
                        "example": "10:00 AM - 12:00 PM"
                    },
                    "items": {
                        "description": "List of items in the order",
                        "type": "array",
                        "items": {
                            "required": [
                                "mealId",
                                "quantity"
                            ],
                            "properties": {
                                "meal_id": {
                                    "description": "The ID of the meal",
                                    "type": "integer",
                                    "example": 1
                                },
                                "quantity": {
                                    "description": "The quantity of the meal",
                                    "type": "integer",
                                    "example": 2
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
...
            "UpdateOrderRequest": {
                "properties": {
                    "status": {
                        "description": "The status of the order",
                        "type": "string",
                        "enum": [
                            "pending",
                            "confirmed",
                            "preparing",
                            "delivered",
                            "cancelled"
                        ]
                    },
                    "deliveryAddress": {
                        "description": "The delivery address for the order",
                        "type": "string",
                        "maxLength": 255
                    },
                    "deliveryTimeSlot": {
                        "description": "The delivery time slot for the order",
                        "type": "string",
                        "maxLength": 50
                    }
                },
                "type": "object"
            },
...
            "OrderResource": {
                "title": "Order Resource",
                "description": "Order resource representation",
                "properties": {
                    "id": {
                        "description": "The ID of the order",
                        "type": "integer"
                    },
                    "user": {
                        "description": "The user who placed the order",
                        "properties": {
                            "id": {
                                "description": "The ID of the user",
                                "type": "integer"
                            },
                            "name": {
                                "description": "The name of the user",
                                "type": "string"
                            }
                        },
                        "type": "object"
                    },
                    "total": {
                        "description": "The total amount of the order in currency",
                        "type": "number",
                        "format": "float"
                    },
                    "status": {
                        "description": "The status of the order",
                        "type": "string"
                    },
                    "deliveryAddress": {
                        "description": "The delivery address of the order",
                        "type": "string"
                    },
                    "deliveryTimeSlot": {
                        "description": "The delivery time slot of the order",
                        "type": "string"
                    },
                    "items": {
                        "description": "The items in the order",
                        "type": "array",
                        "items": {
                            "properties": {
                                "mealId": {
                                    "description": "The ID of the meal",
                                    "type": "integer"
                                },
                                "mealTitle": {
                                    "description": "The title of the meal",
                                    "type": "string"
                                },
                                "quantity": {
                                    "description": "The quantity of the meal",
                                    "type": "integer"
                                },
                                "unit_price": {
                                    "description": "The unit price of the meal in currency",
                                    "type": "number",
                                    "format": "float"
                                },
                                "subtotal": {
                                    "description": "The subtotal of the meal in currency",
                                    "type": "number",
                                    "format": "float"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "placedAt": {
                        "description": "The date and time when the order was placed",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "description": "The date when the order was last updated",
                        "type": "string",
                        "format": "date"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}